Refactor clone code (#509)
* Refactor clone code


* Make urls.ts file and refactor
