+++ src/utils.ts
@@ -6,6 +6,7 @@ import { prompt } from 'inquirer';
 import * as pluralize from 'pluralize';
 import { ClaspToken, DOT, DOTFILE, ProjectSettings } from './dotfile';
 import {Manifest} from './manifest';
+import { URL } from './urls';
 
 const ucfirst = require('ucfirst');
 const isOnline = require('is-online');
@@ -56,18 +57,6 @@ export function getOAuthSettings(local: boolean): Promise<ClaspToken> {
       logError(err, ERROR.NO_CREDENTIALS(local));
     });
 }
-
-// Helpers to get Apps Script project URLs
-export const URL = {
-  APIS: (projectId: string) => `https://console.developers.google.com/apis/dashboard?project=${projectId}`,
-  CREDS: (projectId: string) => `https://console.developers.google.com/apis/credentials?project=${projectId}`,
-  LOGS: (projectId: string) =>
-    `https://console.cloud.google.com/logs/viewer?project=${projectId}&resource=app_script_function`,
-  SCRIPT_API_USER: 'https://script.google.com/home/usersettings',
-  // It is too expensive to get the script URL from the Drive API. (Async/not offline)
-  SCRIPT: (scriptId: string) => `https://script.google.com/d/${scriptId}/edit`,
-  DRIVE: (driveId: string) => `https://drive.google.com/open?id=${driveId}`,
-};
 
 // Error messages (some errors take required params)
 export const ERROR = {
